section 10 clear

+++ src/controllers/videoController.js
@@ -4,7 +4,7 @@ import User from "../models/User";
 
 export const home = async (req, res) => {
   try{
-    const videos = await Video.find({}).sort({createdAt: "desc"});
+    const videos = await Video.find({}).sort({createdAt: "desc"}).populate("owner");
     return res.render("home", {pageTitle: "Home", videos}); 
   }
   catch(error){
@@ -23,19 +23,30 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const {id} = req.params;
+  const {user: {_id}} = req.session;
   const video = await Video.findById(id);
   if(video == null)
     return res.status(404).render("404", {pageTitle: "Video not found,"});
+  
+  console.log(video.owner, _id);
+  if(String(video.owner) != String(_id)){
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", {pageTitle: `Edit: ${video.title}`, video});
 };
 
 export const postEdit = async (req,res) => {
   const {id} = req.params;
+  const {user: {_id}} = req.session;
   const {title, description, hashtags} = req.body;
   const video = await Video.exists({id: id});
   if(!video) 
     return res.status(404).render("404", {pageTitle: "Video not found,"});
   
+  if(String(video.owner) != String(_id)){
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndUpdate(id, {title, description, hashtags: Video.formatHashtags(hashtags)});
 
   return res.redirect(`/videos/${id}`);
@@ -47,7 +58,7 @@ export const getUpload = (req, res) => {
 
 export const postUpload = async (req, res) => {
   const {
-    user: {_id},
+    user: { _id },
   } = req.session;
   const {file} = req;
   const {title, description, hashtags}= req.body;
@@ -69,6 +80,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const {id} = req.params;
+  const {user: {_id}} = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if(!video){
+    return res.status(404).render("404", {pageTitle: "Video not found."});
+  }
+  if(String(video.owner) != String(_id)){
+    return res.status(403).redirect("/");
+  }
+  user.videos.pop(id);
+  await user.save();
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
@@ -81,7 +103,7 @@ export const search = async (req, res) => {
       title: {
         $regex: new RegExp(keyword, "i"),
       },
-    })
+    }).populate("owner");
     return res.render("search", {pageTitle: "Search", videos});
   }
   return res.render("search", {pageTitle: "Search", videos});
